playbook: remove debops, comment out untested playbooks
We remove debops for now; it sucks in a lot of unwanted dependencies
such as mysql-client, it requests packages by default which are in
jessie but not in stretch and it is still based on python 2.7. Ansible
has rules for debian management as well, which might be less sophisti-
cated, but as of know, I don't know which parto of this sophistication
we really need :-)

If we see a need for it, we can re-add it later on.

Signed-off-by: Robert Schwebel <robert@schwebel.de>